third_party/blink/web_tests/external/wpt/editing/run/forwarddelete_7001-last-expected.txt

This is a testharness.js-based test.
Found 23 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","p"],["forwarddelete",""]] "<p><font size=3>foo[]</font><p><font size=5>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font size=\\"3\\">foo</font><font size=\\"5\\">bar</font></p>" but got "<p><font size=\\"3\\">foo</font><span style=\\"font-size:x-large\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","div"],["forwarddelete",""]] "<p><font size=4>foo[]</font><p><font size=5>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font size=\\"4\\">foo</font><font size=\\"5\\">bar</font></p>" but got "<p><font size=\\"4\\">foo</font><span style=\\"font-size:x-large\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","p"],["forwarddelete",""]] "<p><font size=4>foo[]</font><p><font size=5>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font size=\\"4\\">foo</font><font size=\\"5\\">bar</font></p>" but got "<p><font size=\\"4\\">foo</font><span style=\\"font-size:x-large\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","div"],["forwarddelete",""]] "<p><font color=blue>foo[]</font><p><font size=5>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font color=\\"blue\\">foo</font><font size=\\"5\\">bar</font></p>" but got "<p><font color=\\"blue\\">foo</font><span style=\\"font-size:x-large\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","p"],["forwarddelete",""]] "<p><font color=blue>foo[]</font><p><font size=5>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font color=\\"blue\\">foo</font><font size=\\"5\\">bar</font></p>" but got "<p><font color=\\"blue\\">foo</font><span style=\\"font-size:x-large\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","div"],["forwarddelete",""]] "<p><font size=5>foo[]</font><p><font color=blue>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font size=\\"5\\">foo</font><font color=\\"blue\\">bar</font></p>" but got "<p><font size=\\"5\\">foo</font><span style=\\"color:rgb(0, 0, 255)\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","p"],["forwarddelete",""]] "<p><font size=5>foo[]</font><p><font color=blue>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font size=\\"5\\">foo</font><font color=\\"blue\\">bar</font></p>" but got "<p><font size=\\"5\\">foo</font><span style=\\"color:rgb(0, 0, 255)\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","div"],["forwarddelete",""]] "<p><font face=monospace>foo[]</font><p><font face=sans-serif>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font face=\\"monospace\\">foo</font><font face=\\"sans-serif\\">bar</font></p>" but got "<p><font face=\\"monospace\\">foo</font><span style=\\"font-family:sans-serif\\">bar</span></p>"
[FAIL] [["styleWithCSS","false"],["defaultparagraphseparator","p"],["forwarddelete",""]] "<p><font face=monospace>foo[]</font><p><font face=sans-serif>bar</font>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<p><font face=\\"monospace\\">foo</font><font face=\\"sans-serif\\">bar</font></p>" but got "<p><font face=\\"monospace\\">foo</font><span style=\\"font-family:sans-serif\\">bar</span></p>"
[FAIL] [["forwarddelete",""]] "<div><span style=\\"display:list-item\\">abc[]</span><span style=\\"display:list-item\\">def</span></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><span style=\\"display:list-item\\">abcdef</span></div>" but got "<span style=\\"display:list-item\\">abcdef</span>"
[FAIL] [["forwarddelete",""]] "<div>{abc</div><div>def</div>}" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div><br></div>" but got "<br>"
[FAIL] [["forwarddelete",""]] "<table><tr><td>a<td><b>[b</b><td><b>c</b><td><b>d]</b><td>e</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><td>a</td><td><br></td><td><br></td><td><br></td><td>e</td></tr></tbody></table>" but got "<table><tbody><tr><td>a</td><td><br></td><td></td><td></td><td>e</td></tr></tbody></table>"
[FAIL] [["forwarddelete",""]] "<table><tr><th>a<th><b>[b</b><th><b>c</b><th><b>d]</b><th>e</table>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<table><tbody><tr><th>a</th><th><br></th><th><br></th><th><br></th><th>e</th></tr></tbody></table>" but got "<table><tbody><tr><th>a</th><th><br></th><th></th><th></th><th>e</th></tr></tbody></table>"
[FAIL] [["forwarddelete",""]] "012[]<div style=display:inline-flex><span>abc</span><span>def</span></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "012abc<div style=\\"display:inline-flex\\"><span>def</span></div>" but got "012<div style=\\"display:inline-flex\\">abc<span>def</span></div>"
[FAIL] [["forwarddelete",""]] "012[]<div style=display:inline-grid><span>abc</span><span>def</span></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "012abc<div style=\\"display:inline-grid\\"><span>def</span></div>" but got "012<div style=\\"display:inline-grid\\">abc<span>def</span></div>"
[FAIL] [["forwarddelete",""]] "<div style=display:grid><span>{}<br></span></div><div>abc</div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div>abc</div>" but got "<div style=\\"display:grid\\"><span>abc</span></div>"
[FAIL] [["forwarddelete",""]] "<div style=display:flex><span>abc[]</span><span style=display:contents>def</span></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"display:flex\\"><span>abcdef</span></div>" but got "<div style=\\"display:flex\\"><span>abcdef</span><span style=\\"display:contents\\"><br></span></div>"
[FAIL] [["forwarddelete",""]] "<div style=display:flex><span style=display:contents>abc[]</span><span>def</span></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"display:flex\\"><span style=\\"display:contents\\">abcdef</span></div>" but got "<div style=\\"display:flex\\"><span style=\\"display:contents\\">abc</span>def</div>"
[FAIL] [["forwarddelete",""]] "<div style=display:grid><span>abc[]</span><span style=display:contents>def</span></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"display:grid\\"><span>abcdef</span></div>" but got "<div style=\\"display:grid\\"><span>abcdef</span><span style=\\"display:contents\\"><br></span></div>"
[FAIL] [["forwarddelete",""]] "<div style=display:grid><span style=display:contents>abc[]</span><span>def</span></div>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "<div style=\\"display:grid\\"><span style=\\"display:contents\\">abcdef</span></div>" but got "<div style=\\"display:grid\\"><span style=\\"display:contents\\">abc</span>def</div>"
[FAIL] [["forwarddelete",""]] "abc[]<object>def</object>" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "abc<object>ef</object>" but got "abc<object>def</object>"
[FAIL] [["forwarddelete",""]] "abc[]<meter>def</meter>ghi" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "abcghi" but got "abc<meter></meter>ghi"
[FAIL] [["forwarddelete",""]] "abc[]<audio>def</audio>ghi" compare innerHTML
  assert_equals: Unexpected innerHTML (after normalizing inline style) expected "abchi" but got "abcghi"
Harness: the test ran to completion.
